Nest interfaces and types under a WoWAPI namespace
Declared interfaces and types are no longer available in the global namespace,
instead, they can be referenced through the WoWAPI-namespace, e.g. WoWAPI.Frame.

Added an interface for Object and moved some method signatures between Object
and UIObject to comply with the widget hierarchy defined here:
https://github.com/Ketho/BlizzardInterfaceResources/blob/live/Resources/WidgetAPI.lua

Corrected a typo in AchievementLink and CombatRatingIdentifier.